+++ open_connection.go
@@ -1,8 +1,6 @@
 package main
 
-import (
-	"log"
-)
+import "log"
 
 type ViewMode string
